Several improvements due to trying to use it in a real project

+++ app/src/main/kotlin/com/github/silasgermany/database/sql/BaseTable.kt
@@ -5,5 +5,5 @@ import com.github.silasgermany.complexormapi.ComplexOrmTable
 
 open class BaseTable(initMap: MutableMap<String, Any?> = default): ComplexOrmTable(initMap) {
     @ComplexOrmReadAlways
-    var ownId: Int by initMap
+    var ownId: Long by initMap
 }
